add hero section
